Remove PySharedPtr in favor of pybind11 smart_holder branch.
PySharedPtr didn't work at all how its authors thought it did - it just
unconditionally leaked any trampoline-wrapped object passed through
a shared_ptr to Python.

The fix for this requires changes to pybind11 itself, but they're
upstream in the long-lived smart_holder branch (which comes with a
number of other nice features).  We now vendor that branch as a EUPS
package.